CONFIG: Update Api Acceptance Tests
Closes #23
