+++ TSwim.Core.API.Tests.Acceptance/Brokers/ApiBroker.Home.cs
@@ -5,10 +5,16 @@
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 // -----------------------------------------------------------------------------------
 
-using Xunit;
+using System.Threading.Tasks;
 
-namespace TSwim.Core.Tests.Acceptance.Brokers
+namespace TSwim.Core.Api.Tests.Acceptance.Brokers
 {
-    [CollectionDefinition(nameof(ApiTestCollection))]
-    public class ApiTestCollection : ICollectionFixture<ApiBroker> { }
+    public partial class ApiBroker
+    {
+        private const string HomeRelativeUrl = "api/home";
+
+        public async ValueTask<string> GetHomeMessage() =>
+            await this.apiFactoryClient.GetContentStringAsync(HomeRelativeUrl);
+
+    }
 }
