--- TSwim.Core/Brokers/Storages/StorageBroker.cs
@@ -1,32 +1,0 @@
-﻿// -----------------------------------------------------------------------------------
-// Participation to the Coalition of the Good-Hearted Engineers WorldWide
-// Copyright © TSwim and contributors. All rights reserved.
-// GLHays ©2022 Founder,Developer - (TS)Warehouse Inventory Management
-// Licensed under the MIT license. See LICENSE file in the project root for details.
-// -----------------------------------------------------------------------------------
-
-using EFxceptions;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-
-namespace TSwim.Core.Brokers.Storages
-{
-    public partial class StorageBroker : EFxceptionsContext, IStorageBroker
-    {
-        private readonly IConfiguration configuration;
-
-        public StorageBroker(IConfiguration configuration)
-        {
-            this.configuration = configuration;
-            this.Database.Migrate();
-        }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            string connectionString = this.configuration
-                .GetConnectionString(name: "DefaultConnection");
-
-            optionsBuilder.UseSqlServer(connectionString);
-        }
-    }
-}
