--- TSwim.Core/Controllers/HomeController.cs
@@ -1,21 +1,0 @@
-﻿// -----------------------------------------------------------------------------------
-// Participation to the Coalition of the Good-Hearted Engineers WorldWide
-// Copyright © TSwim and contributors. All rights reserved.
-// GLHays ©2022 Founder,Developer - (TS)Warehouse Inventory Management
-// Licensed under the MIT license. See LICENSE file in the project root for details.
-// -----------------------------------------------------------------------------------
-
-using Microsoft.AspNetCore.Mvc;
-using RESTFulSense.Controllers;
-
-namespace TSwim.Core.Controllers
-{
-    [ApiController]
-    [Route("api/[controller]")]
-    public class HomeController : RESTFulController
-    {
-        [HttpGet]
-        public ActionResult<string> Get() =>
-            Ok("Thank you Mario! But the princess is in another castle.");
-    }
-}
