--- TSwim.Core/Program.cs
@@ -1,23 +1,0 @@
-// -----------------------------------------------------------------------------------
-// Participation to the Coalition of the Good-Hearted Engineers WorldWide
-// Copyright © TSwim and contributors. All rights reserved.
-// GLHays ©2022 Founder,Developer - (TS)Warehouse Inventory Management
-// Licensed under the MIT license. See LICENSE file in the project root for details.
-// -----------------------------------------------------------------------------------
-
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
-
-namespace TSwim.Core
-{
-    public class Program
-    {
-        public static void Main(string[] args) =>
-            CreateHostBuilder(args).Build().Run();
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                webBuilder.UseStartup<Startup>());
-    }
-}
