--- TSwim.Core/Startup.cs
@@ -1,67 +1,0 @@
-// -----------------------------------------------------------------------------------
-// Participation to the Coalition of the Good-Hearted Engineers WorldWide
-// Copyright © TSwim and contributors. All rights reserved.
-// GLHays ©2022 Founder,Developer - (TS)Warehouse Inventory Management
-// Licensed under the MIT license. See LICENSE file in the project root for details.
-// -----------------------------------------------------------------------------------
-
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.OpenApi.Models;
-using TSwim.Core.Brokers.Storages;
-
-namespace TSwim.Core
-{
-    public class Startup
-    {
-        public Startup(IConfiguration configuration) =>
-            Configuration = configuration;
-
-        public IConfiguration Configuration { get; }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddControllers();
-            services.AddDbContext<StorageBroker>();
-            services.AddTransient<IStorageBroker, StorageBroker>();
-
-            services.AddSwaggerGen(options =>
-            {
-                var openApiInfo = new OpenApiInfo
-                {
-                    Title = "TSwim.Core",
-                    Version = "v1"
-                };
-
-                options.SwaggerDoc(
-                    name: "v1",
-                    info: openApiInfo);
-            });
-        }
-
-        public void Configure(IApplicationBuilder applicationBuilder, IWebHostEnvironment webHostEnvironment)
-        {
-            if (webHostEnvironment.IsDevelopment())
-            {
-                applicationBuilder.UseDeveloperExceptionPage();
-                applicationBuilder.UseSwagger();
-
-                applicationBuilder.UseSwaggerUI(options =>
-                {
-                    options.SwaggerEndpoint(
-                        url: "/swagger/v1/swagger.json",
-                        name: "TSwim.Core v1");
-                });
-            }
-
-            applicationBuilder.UseHttpsRedirection();
-            applicationBuilder.UseRouting();
-            applicationBuilder.UseAuthorization();
-
-            applicationBuilder.UseEndpoints(endpoints => endpoints.MapControllers());
-        }
-    }
-}
