--- TSwim.Core/appsettings.json
@@ -1,15 +1,0 @@
-// Be aware that the ConnectionString variable here is for code clarity.
-// DO NOT INCLUDE SENSITIVE DATA LIKE USERNAMES AND PASSWORDS AND EXPOSE TO THE PUBLIC
-{
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=(localdb)\\MSSQLLocalDB;Database=TSwimDB;Trusted_Connection=True;MultipleActiveResultSets=true"
-  },
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "AllowedHosts": "*"
-}
